Conflicts:
	phenoscapeII/src/outputter/ELKReasoner.java
	phenoscapeII/src/outputter/EQPerformanceEvaluation.java
	phenoscapeII/src/outputter/XML2EQ.java

+++ phenoscapeII/src/outputter/XML2EQ.java
@@ -21,6 +21,27 @@ import org.jdom.Element;
 import org.jdom.input.SAXBuilder;
 import org.jdom.xpath.XPath;
 
+import outputter.data.CompositeEntity;
+import outputter.data.CompositeQuality;
+import outputter.data.EQProposals;
+import outputter.data.Entity;
+import outputter.data.EntityProposals;
+import outputter.data.FormalConcept;
+import outputter.data.NegatedQuality;
+import outputter.data.Quality;
+import outputter.data.QualityProposals;
+import outputter.data.RelationalQuality;
+import outputter.data.SimpleEntity;
+import outputter.knowledge.Dictionary;
+import outputter.knowledge.ELKReasoner;
+import outputter.knowledge.TermOutputerUtilities;
+import outputter.output.HTMLOutput;
+import outputter.prep.XMLNormalizer;
+import outputter.process.BinaryCharacterStatementParser;
+import outputter.process.CharacterStatementParser;
+import outputter.process.StateStatementParser;
+import outputter.search.TermSearcher;
+
 
 
 /* annotation guideline: http://phenoscape.org/wiki/Guide_to_Character_Annotation */
@@ -74,6 +95,7 @@ public class XML2EQ {
 
 	public static TermOutputerUtilities ontoutil = new TermOutputerUtilities();
 	public static ELKReasoner elk; 
+	private static boolean recordperformance = true;
 	static{
 		try{
 			//TODO: figure out why the two calls give different results?
@@ -114,27 +136,32 @@ public class XML2EQ {
 		this.glosstable = glosstable;
 		//this.keyentities = new ArrayList<Hashtable<String,String>>();
 
+		if(isRecordperformance()){
+			if(dictionary.conn == null){
+				Class.forName("com.mysql.jdbc.Driver");
+				dictionary.conn = DriverManager.getConnection(ApplicationUtilities.getProperty("database.url"));
+			}
 
-		if(dictionary.conn == null){
-			Class.forName("com.mysql.jdbc.Driver");
-			dictionary.conn = DriverManager.getConnection(ApplicationUtilities.getProperty("database.url"));
+
+			Statement stmt = dictionary.conn.createStatement();
+			// label and id fields are ontology-related fields
+			// other fields are raw text
+			// entity and quality fields are atomic
+			// qualitynegated fields are alternative to quality and is composed as "not quality" for qualitynegated, "not(quality)" for qualitynegatedlabel, the "quality" has id
+			// qualityid
+			// qualitymodifier/label/id and entitylocator/label/id may hold multiple values separated by "," which preserves the order of multiple values
+			stmt.execute("drop table if exists " + outputtable);
+			System.out.println("create table if not exists " + outputtable
+					+ " (id int(11) not null unique auto_increment primary key, source varchar(500), characterID varchar(100), characterlabel varchar(1000), stateID varchar(100), statelabel text, "
+					+ " entity varchar(3000),entitylabel varchar(3000), entityid varchar(3000), " + "quality varchar(3000),qualitylabel varchar(3000), qualityid varchar(3000),"+"relatedentity varchar(3000),relatedentitylabel varchar(3000), relatedentityid varchar(3000))");
+			//stmt.execute("create table if not exists " + outputtable
+			//		+ " (id int(11) not null unique auto_increment primary key, source varchar(500), characterID varchar(100), characterlabel varchar(1000), stateID varchar(100), statelabel text, "
+			//		+ " entity varchar(3000),entitylabel varchar(3000), entityid varchar(3000), " + "quality varchar(3000),qualitylabel varchar(3000), qualityid varchar(3000),"+"relatedentity varchar(3000),relatedentitylabel varchar(3000), relatedentityid varchar(3000))" );
+			stmt.execute("create table if not exists " + outputtable
+					+ " (id int(11) not null unique auto_increment primary key, source varchar(500), characterID varchar(100), characterlabel varchar(1000), stateID varchar(100), statelabel text, "
+					+ " entity text,entitylabel text, entityid text, " + "quality text,qualitylabel text, qualityid text,"+"relatedentity text,relatedentitylabel text, relatedentityid text)" );
+
 		}
-
-		Statement stmt = dictionary.conn.createStatement();
-		// label and id fields are ontology-related fields
-		// other fields are raw text
-		// entity and quality fields are atomic
-		// qualitynegated fields are alternative to quality and is composed as "not quality" for qualitynegated, "not(quality)" for qualitynegatedlabel, the "quality" has id
-		// qualityid
-		// qualitymodifier/label/id and entitylocator/label/id may hold multiple values separated by "," which preserves the order of multiple values
-		stmt.execute("drop table if exists " + outputtable);
-		System.out.println("create table if not exists " + outputtable
-				+ " (id int(11) not null unique auto_increment primary key, source varchar(500), characterID varchar(100), characterlabel varchar(1000), stateID varchar(100), statelabel text, "
-				+ " entity varchar(3000),entitylabel varchar(3000), entityid varchar(3000), " + "quality varchar(3000),qualitylabel varchar(3000), qualityid varchar(3000),"+"relatedentity varchar(3000),relatedentitylabel varchar(3000), relatedentityid varchar(3000))");
-		stmt.execute("create table if not exists " + outputtable
-				+ " (id int(11) not null unique auto_increment primary key, source varchar(500), characterID varchar(100), characterlabel varchar(1000), stateID varchar(100), statelabel text, "
-				+ " entity varchar(3000),entitylabel varchar(3000), entityid varchar(3000), " + "quality varchar(3000),qualitylabel varchar(3000), qualityid varchar(3000),"+"relatedentity varchar(3000),relatedentitylabel varchar(3000), relatedentityid varchar(3000))" );
-
 		pathStructure = XPath.newInstance(".//structure");
 		pathWholeOrgStrucChar= XPath.newInstance(".//structure[@name='"+ApplicationUtilities.getProperty("unknown.structure.name")+"']/character");
 		pathCharacter = XPath.newInstance(".//character");
@@ -192,8 +219,12 @@ public class XML2EQ {
 				LOGGER.error("", e);
 			}
 		}
-		HTMLOutput output = new HTMLOutput();
-		output.outputHTML(this.outputtable,"curator",0);
+
+		if(isRecordperformance()){
+			HTMLOutput output = new HTMLOutput();
+			output.outputHTML(this.outputtable,"curator",0);
+		}
+
 
 		elk.dispose();
 	}
@@ -237,15 +268,15 @@ public class XML2EQ {
 
 		//for (EQStatementProposals EQ : allEQs) {
 		for (EQProposals EQ : allEQs) {			
-		this.insertEQs2Table(EQ);
-		//	System.out.println(EQ.toString());
+			if(isRecordperformance()) this.insertEQs2Table(EQ);
+			System.out.println(EQ.toString());
 		}
 
 	}
 
-	
-	
-	
+
+
+
 	private void insertEQs2Table(EQProposals eQ)
 	{
 		String entity ="";
@@ -258,7 +289,7 @@ public class XML2EQ {
 		String relatedentitylabel="";
 		String relatedentityid="";
 		//Read all Entity Proposals and store as comma separated values
-		for(Entity e:eQ.entity.getProposals())
+		for(Entity e: eQ.getEntity().getProposals())
 		{
 			entitylabel+=e.getLabel()+" Score:["+e.getConfidienceScore()+"]@,";
 			if(e instanceof CompositeEntity)
@@ -272,16 +303,16 @@ public class XML2EQ {
 				entityid+=e.getId()+" Score:["+e.getConfidienceScore()+"]@,";
 			}
 		}
-		
+
 		entity = entity.replaceAll("(@,)$", "");
 		entitylabel = entitylabel.replaceAll("(@,)$", "");
 		entityid = entityid.replaceAll("(@,)$", "");
 		entity =sort(entity);
 		entitylabel =sort(entitylabel);
 		entityid =sort(entityid);
-		
+
 		//Read all Quality Proposals and store as comma separated values
-		for(Quality q:eQ.quality.getProposals())
+		for(Quality q:eQ.getQuality().getProposals())
 		{
 			if(q instanceof RelationalQuality)
 			{
@@ -290,7 +321,7 @@ public class XML2EQ {
 				qualityid+=q.getId()+" Score:["+q.getConfidienceScore()+"]@,";
 				qualitylabel+=q.getLabel()+" Score:["+q.getConfidienceScore()+"]@,";
 				//Reading all related entities and store as comma separated values
-				for(Entity e:((RelationalQuality) q).relatedentity.getProposals())
+				for(Entity e:((RelationalQuality) q).getRelatedentity().getProposals())
 				{
 					relatedentitylabel+=e.getLabel()+" Score:["+e.getConfidienceScore()+"]@,";
 					if(e instanceof CompositeEntity)
@@ -325,34 +356,33 @@ public class XML2EQ {
 				qualitylabel+=q.getLabel()+" Score:["+q.getConfidienceScore()+"]@,";
 			}
 		}
-		
+
 		relatedentity = relatedentity.replaceAll("(@,)$", "");
 		relatedentitylabel = relatedentitylabel.replaceAll("(@,)$", "");
 		relatedentityid = relatedentityid.replaceAll("(@,)$", "");
 		quality = quality.replaceAll("(@,)$", "");
 		qualitylabel = qualitylabel.replaceAll("(@,)$", "");
 		qualityid = qualityid.replaceAll("(@,)$", "");
-		
+
 		quality = sort(quality);
 		qualitylabel = sort(qualitylabel);
 		qualityid = sort(qualityid);
-		
+
 		relatedentity = sort(relatedentity);
 		relatedentitylabel = sort(relatedentitylabel);
 		relatedentityid = sort(relatedentityid);
-		
-		
+
+
 		String sql = "insert into "+this.outputtable +" (source,characterID,characterlabel,stateID,statelabel, entity,"+
 					 "entitylabel,entityid,quality,qualitylabel,qualityid,relatedentity,relatedentitylabel,relatedentityid) values"+
 					 "(?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
-		
 		try {
 			PreparedStatement preparedStatement = dictionary.conn.prepareStatement(sql);
-			preparedStatement.setString(1, eQ.sourceFile);
-			preparedStatement.setString(2, eQ.characterId);
-			preparedStatement.setString(3,eQ.characterlabel);
-			preparedStatement.setString(4,eQ.stateId);
-			preparedStatement.setString(5,eQ.description);
+			preparedStatement.setString(1, eQ.getSourceFile());
+			preparedStatement.setString(2, eQ.getCharacterId());
+			preparedStatement.setString(3,eQ.getCharacterlabel());
+			preparedStatement.setString(4,eQ.getStateId());
+			preparedStatement.setString(5,eQ.getDescription());
 			preparedStatement.setString(6,entity);
 			preparedStatement.setString(7,entitylabel);
 			preparedStatement.setString(8,entityid);
@@ -363,101 +393,96 @@ public class XML2EQ {
 			preparedStatement.setString(13,relatedentitylabel);
 			preparedStatement.setString(14,relatedentityid);
 			preparedStatement.executeUpdate();
-			
+
 		} catch (SQLException e1) {
 			e1.printStackTrace();
 		}
-		
-		
+
+
 	}
 
-	/*
-	 * 
-	 * sort the proposals according to the confidence scores
-	 * 
-	 */
-private String sort(String groups) 
-{
-
-	String individualstrings[] = groups.split("(@,)");//Splitting original Proposals into individual tokens upon ],(Comma alone cannot be used, so using ],)
-	
-
-	String sortedstring ="";
-	
-
-	Hashtable <Float,ArrayList<String>> holder = new Hashtable <Float,ArrayList<String>>();
-	//Extracting the float value and creating an has Float, String
-	for(String token:individualstrings)
+	private String sort(String groups) 
 	{
-		if((token!=null) &&(token.equals("")==false))
+
+		String individualstrings[] = groups.split("(@,)");//Splitting original Proposals into individual tokens upon ],(Comma alone cannot be used, so using ],)
+
+
+		String sortedstring ="";
+
+
+		Hashtable <Float,ArrayList<String>> holder = new Hashtable <Float,ArrayList<String>>();
+		//Extracting the float value and creating an has Float, String
+		for(String token:individualstrings)
 		{
-		Float value = Float.parseFloat(token.split("(\\[)")[1].replaceAll("]$", ""));
-		
-		if(holder.get(value)==null)
-		{
-			ArrayList<String> list = new ArrayList<String>();
-			list.add(token);
-			holder.put(value, list);
-		}
-		else
-		{
-			ArrayList<String> list = holder.get(value);
-			list.add(token);
-			holder.put(value, list);
-		}
-	}
-	}
-	//sorting the Keyset of the hashtable
-	Set<Float> keys =holder.keySet();
-	Object keyarray[] = keys.toArray();
-	ArrayList<Float> sortedkeys = new ArrayList<Float>();
-	
-	for(int i=0;i<keyarray.length;i++)
-	{
-		for(int j=i+1;j<keyarray.length;j++)
-		{
-			if((Float)keyarray[i]<(Float)keyarray[j])
+			if((token!=null) &&(token.equals("")==false))
 			{
-				Float temp = (Float) keyarray[i];
-				keyarray[i] = keyarray[j];
-				keyarray[j] = temp;
+			Float value = Float.parseFloat(token.split("(\\[)")[1].replaceAll("]$", ""));
+
+			if(holder.get(value)==null)
+			{
+				ArrayList<String> list = new ArrayList<String>();
+				list.add(token);
+				holder.put(value, list);
+			}
+			else
+			{
+				ArrayList<String> list = holder.get(value);
+				list.add(token);
+				holder.put(value, list);
 			}
 		}
-	}
-	
-	//Iterating through the sorted keyset and creating the sorted final String
-	
-	for(int i=0;i<keyarray.length;i++)
-	{
-		ArrayList<String> templist = holder.get(keyarray[i]);
-		Collections.sort(templist);
-		
-		for(String temp:templist)
-		{
-			sortedstring+=temp+"@,";
 		}
-	}
-	return sortedstring.replaceAll("(@,)$", "");
-	
-}
+		//sorting the Keyset of the hashtable
+		Set<Float> keys =holder.keySet();
+		Object keyarray[] = keys.toArray();
+		ArrayList<Float> sortedkeys = new ArrayList<Float>();
 
-/*	private ArrayList<String> printEntity(ArrayList<Entity> proposals) {
+		for(int i=0;i<keyarray.length;i++)
+		{
+			for(int j=i+1;j<keyarray.length;j++)
+			{
+				if((Float)keyarray[i]<(Float)keyarray[j])
+				{
+					Float temp = (Float) keyarray[i];
+					keyarray[i] = keyarray[j];
+					keyarray[j] = temp;
+				}
+			}
+		}
+
+		//Iterating through the sorted keyset and creating the sorted final String
+
+		for(int i=0;i<keyarray.length;i++)
+		{
+			ArrayList<String> templist = holder.get(keyarray[i]);
+			Collections.sort(templist);
+
+			for(String temp:templist)
+			{
+				sortedstring+=temp+"@,";
+			}
+		}
+		return sortedstring.replaceAll("(@,)$", "");
+
+	}
+
+	/*	private ArrayList<String> printEntity(ArrayList<Entity> proposals) {
 
 		ArrayList<String> entitystrings = new ArrayList<String>();
 		String entity ="";
 		String entitylabel="";
 		String entityid="";
-		
+
 		for(Entity e:proposals)
 		{
 			if(e instanceof CompositeEntity)
 			{
 				ArrayList<String> entitystring = printEntity(((CompositeEntity) e).getEntities());
-				
+
 				entity+=entitystring.get(0);
 				entitylabel+=entitystring.get(1);
 				entityid+=entitystring.get(2);
-				
+
 			}
 			else{
 				entity+=e.getString()+",";
@@ -465,11 +490,11 @@ private String sort(String groups)
 				entityid+=e.getId()+",";
 			}
 		}
-		
+
 		entitystrings.add(entity.replaceAll(",$",""));
 		entitystrings.add(entitylabel.replaceAll(",$",""));
 		entitystrings.add(entityid.replaceAll(",$",""));
-		
+
 		return entitystrings;
 	}*/
 
@@ -929,31 +954,31 @@ private String sort(String groups)
 						ArrayList<FormalConcept> qs =  new TermSearcher().searchTerm(ngram, "quality"); 
 						if(qs!=null){
 							for(FormalConcept fc: qs){
-							String qlabel = fc.getLabel();
-							String cp = commonParent(qlabel, qualitylabels);
-							if(cp!=null && cp.matches(".*?\\b("+dictionary.patoupperclasses+")\\b.*")){//TODO matches parent quality or any of its offsprings is fine.
-								//EQStatementProposals EQp = relatedEQ(stateid, ngram);
-								EQProposals EQp = relatedEQ(stateid, ngram);
-								if(EQp==null){ //add one
-									//EQp = new EQStatementProposals();
-									EQp = new EQProposals();
-									//EQStatement EQ = new EQStatement();
-									//add metadata
-									String characterid = "";
-									try{
-										Element statement = (Element) XPath.selectSingleNode(root, ".//statement[@state_id='"+stateid+"']");
-										characterid = statement.getAttributeValue("character_id");
-									}catch(Exception e){
-										LOGGER.error("", e);
+								String qlabel = fc.getLabel();
+								String cp = commonParent(qlabel, qualitylabels);
+								if(cp!=null && cp.matches(".*?\\b("+dictionary.patoupperclasses+")\\b.*")){//TODO matches parent quality or any of its offsprings is fine.
+									//EQStatementProposals EQp = relatedEQ(stateid, ngram);
+									EQProposals EQp = relatedEQ(stateid, ngram);
+									if(EQp==null){ //add one
+										//EQp = new EQStatementProposals();
+										EQp = new EQProposals();
+										//EQStatement EQ = new EQStatement();
+										//add metadata
+										String characterid = "";
+										try{
+											Element statement = (Element) XPath.selectSingleNode(root, ".//statement[@state_id='"+stateid+"']");
+											characterid = statement.getAttributeValue("character_id");
+										}catch(Exception e){
+											LOGGER.error("", e);
+										}
+										EQp.setSource(src);
+										EQp.setCharacterId(characterid);
+										EQp.setStateId(stateid);
+										EQp.setDescription(text);
+										//EQp.add(EQ);
+										allEQs.add(EQp);
 									}
-									EQp.setSource(src);
-									EQp.setCharacterId(characterid);
-									EQp.setStateId(stateid);
-									EQp.setDescription(text);
-									//EQp.add(EQ);
-									allEQs.add(EQp);
 								}
-							}
 								//accept this result for this stateid
 								/*EQStatement EQ = EQp.getProposals().get(0); //assuming there is only one candidate???
 								EQ.setEntity(keyEQ.getEntity());
@@ -1203,9 +1228,9 @@ private String sort(String groups)
 					if(aEQp.getQuality()!=null){
 						for(Quality Q: aEQp.getQuality().getProposals()){
 							String q = Q!=null?Q.getLabel():""; //ternary operator added => Hariharan
-	
+
 							if(q==null) q="";
-	
+
 							if(q.length()>0) hasquality = true;
 						}
 					}
@@ -1340,7 +1365,7 @@ private String sort(String groups)
 	 * @param entitylabel
 	 */
 	//private void inheritEntityLocator(EQStatement EQ, String entity){
-		private void inheritEntityLocator(EQProposals EQ, String entity){
+	private void inheritEntityLocator(EQProposals EQ, String entity){
 
 		/*	String elid = EQ.get("entitylocatorid");
 		for(Entity keyentity: this.keyentities){
@@ -1424,6 +1449,14 @@ private String sort(String groups)
 		}
 
 
+	}
+
+	public static boolean isRecordperformance() {
+		return recordperformance;
+	}
+
+	public static void setRecordperformance(boolean recordperformance) {
+		XML2EQ.recordperformance = recordperformance;
 	}
 
 }
